refactor: reorganize settings layout for better UX
