+++ src/features/settings/SettingsPage.tsx
@@ -131,10 +131,7 @@ export default function SettingsPage() {
     return {
       general:
         formState.uiLanguage !== globalSettings.uiLanguage ||
-        formState.system.skillsDirectory !==
-          globalSettings.system.skillsDirectory ||
-        formState.system.maxFileUploadSizeMB !==
-          globalSettings.system.maxFileUploadSizeMB,
+        !equal(formState.display, globalSettings.display),
       'ai-models': !equal(
         {
           serviceConfigs: formState.serviceConfigs,
@@ -162,7 +159,6 @@ export default function SettingsPage() {
           windowSize: formState.windowSize,
           maxInputContext: formState.maxInputContext,
           toolCallGroupVisibleCount: formState.toolCallGroupVisibleCount,
-          display: formState.display,
           diffContextLines: formDiffContextLines,
         },
         {
@@ -170,12 +166,13 @@ export default function SettingsPage() {
           windowSize: globalSettings.windowSize,
           maxInputContext: globalSettings.maxInputContext,
           toolCallGroupVisibleCount: globalSettings.toolCallGroupVisibleCount,
-          display: globalSettings.display,
           diffContextLines: globalDiffContextLines,
         },
       ),
       system: !equal(
         {
+          skillsDirectory: formState.system.skillsDirectory,
+          maxFileUploadSizeMB: formState.system.maxFileUploadSizeMB,
           searchIndexFrequencyMinutes:
             formState.system.searchIndexFrequencyMinutes,
           webActionTimeoutSeconds: formState.system.webActionTimeoutSeconds,
@@ -189,6 +186,8 @@ export default function SettingsPage() {
           httpServerExpose: formState.system.httpServerExpose,
         },
         {
+          skillsDirectory: globalSettings.system.skillsDirectory,
+          maxFileUploadSizeMB: globalSettings.system.maxFileUploadSizeMB,
           searchIndexFrequencyMinutes:
             globalSettings.system.searchIndexFrequencyMinutes,
           webActionTimeoutSeconds:
@@ -477,21 +476,9 @@ export default function SettingsPage() {
     (count: number) => update('toolCallGroupVisibleCount', count),
     [update],
   );
-  const handleAgentHubUrlChange = useCallback(
-    (url: string) => update('agentHubUrl', url),
-    [update],
-  );
   const handleLanguageChange = useCallback(
     (lang: string) => update('uiLanguage', lang),
     [update],
-  );
-  const handleSkillsDirectoryChange = useCallback(
-    (path: string) => updateSystem('skillsDirectory', path),
-    [updateSystem],
-  );
-  const handleMaxFileUploadSizeChange = useCallback(
-    (value: number) => updateSystem('maxFileUploadSizeMB', value),
-    [updateSystem],
   );
   const handleMaxRetriesChange = useCallback(
     (value: number) => updateAdvanced('maxRetries', value),
@@ -651,10 +638,8 @@ export default function SettingsPage() {
               <GeneralTab
                 localLanguage={formState.uiLanguage}
                 onChange={handleLanguageChange}
-                skillsDirectory={formState.system.skillsDirectory}
-                onSkillsDirectoryChange={handleSkillsDirectoryChange}
-                localMaxFileUploadSizeMB={formState.system.maxFileUploadSizeMB}
-                onMaxFileUploadSizeChange={handleMaxFileUploadSizeChange}
+                localDisplay={formState.display}
+                onDisplaySettingsChange={updateDisplay}
               />
             </TabsContent>
 
@@ -664,7 +649,6 @@ export default function SettingsPage() {
                 providerEntries={providerEntries}
                 localPreferredModel={formState.preferredModel}
                 localFallbackModel={formState.fallbackModel}
-                localAgentHubUrl={formState.agentHubUrl || ''}
                 localMaxRetries={formState.advanced.maxRetries}
                 localRetryDelay={formState.advanced.retryDelay}
                 localDefaultMaxOutputTokens={
@@ -673,7 +657,6 @@ export default function SettingsPage() {
                 onPendingChange={handlePendingChange}
                 onPreferredModelChange={handlePreferredModelChange}
                 onFallbackModelChange={handleFallbackModelChange}
-                onAgentHubUrlChange={handleAgentHubUrlChange}
                 onMaxRetriesChange={handleMaxRetriesChange}
                 onRetryDelayChange={handleRetryDelayChange}
                 onDefaultMaxOutputTokensChange={
@@ -691,7 +674,6 @@ export default function SettingsPage() {
                   formState.toolCallGroupVisibleCount
                 }
                 localAdvancedSettings={formState.advanced}
-                localDisplay={formState.display}
                 onContextStrategyChange={handleContextStrategyChange}
                 onWindowSizeChange={handleWindowSizeChange}
                 onMaxInputContextChange={handleMaxInputContextChange}
@@ -699,7 +681,6 @@ export default function SettingsPage() {
                   handleToolCallGroupVisibleCountChange
                 }
                 onAdvancedSettingsChange={updateAdvanced}
-                onDisplaySettingsChange={updateDisplay}
               />
             </TabsContent>
